update / (and some other changes)
