+++ src/components/header.astro
@@ -5,10 +5,9 @@ import { cn } from "@/utils/cn";
 import { removeTrailingSlash } from "@/utils/remove-trailing-slash";
 import { FileText, TerminalSquare } from "lucide-astro";
 
-const logoSize = 30;
+const logoSize = 26;
 const iconSize = 16;
 
-// see https://github.com/withastro/astro/issues/5248
 const { pathname } = Astro.url;
 const path = removeTrailingSlash(pathname);
 ---
