+++ src/layouts/main-layout.astro
@@ -10,7 +10,7 @@ const { className } = Astro.props;
 
 <main
   class={cn(
-    "mx-auto w-full max-w-[1300px] px-4 pb-8 pt-[calc(theme(spacing.header)+theme(spacing.8))]",
+    "px-3 pb-12 pt-[calc(theme(spacing.header)+theme(spacing.10))]",
     className,
   )}
 >
