+++ src/layouts/primitive-layout.astro
@@ -12,6 +12,7 @@ interface Props {
 }
 
 const { title, description } = Astro.props;
+
 const canonicalUrl = removeTrailingSlash(
   new URL(Astro.url.pathname, Astro.site).toString(),
 );
