+++ src/pages/index.astro
@@ -1,66 +1,62 @@
 ---
-import { LastFmRecent } from "@/components/react/lastfm-recent";
-import { LastFmStats } from "@/components/react/lastfm-stats";
-import { LastFmWeekly } from "@/components/react/lastfm-weekly";
+import { LastFm } from "@/components/react/lastfm";
 import { Link } from "@/components/react/link";
-import HtmlLayout from "@/layouts/html-layout.astro";
+import MainLayout from "@/layouts/main-layout.astro";
+import PrimitiveLayout from "@/layouts/primitive-layout.astro";
 ---
 
-<HtmlLayout
+<PrimitiveLayout
   title="Charles Wang"
   description="Computer graphics student at the University of Pennsylvania."
 >
-  <main
-    class="flex min-h-[calc(100vh-theme(spacing.footer))] items-center justify-center pt-header"
-  >
+  <MainLayout className="space-y-12">
     <section
-      class="mx-4 my-8 max-w-prose space-y-16 xl:mx-0 xl:my-0 xl:grid xl:max-w-fit xl:grid-flow-col xl:auto-rows-fr xl:grid-cols-8 xl:grid-rows-4 xl:gap-3 xl:space-y-0"
+      class="max-w-content mx-auto animate-fade space-y-[1.15rem] [--delay:100ms]"
     >
-      <div
-        class="animate-fade space-y-4 rounded-xl bg-sweater-10 [--delay:100ms] xl:col-span-4 xl:row-span-4 xl:flex xl:h-[500px] xl:w-[500px] xl:flex-col xl:justify-between xl:bg-sweater-9 xl:p-16"
-      >
-        <h4 class="font-bold text-sweater-2 xl:text-sm">About</h4>
-        <p>
-          Hi, I'm Charles. I'm a student at the University of Pennsylvania
-          majoring in
-          <Link
-            href="http://cg.cis.upenn.edu/dmd.html"
-            newTab
-          >
-            <>Digital Media Design</>
-          </Link>, an interdisciplinary program that combines computer science,
-          graphics, and design.
-        </p>
-        <p>
-          I'm interested in making games and building websites. In my free time
-          I like running, designing album covers, obsessing over metro systems,
-          learning bass guitar, and ricing my Arch setup.
-        </p>
-        <p>Thanks for stopping by!</p>
-        <p>(〜￣▽￣)〜</p>
-      </div>
-
-      <div
-        class="animate-fade rounded-xl bg-sweater-10 [--delay:300ms] xl:col-span-4 xl:row-span-2 xl:flex xl:h-[244px] xl:w-[500px] xl:flex-col xl:justify-between xl:gap-3 xl:bg-sweater-9 xl:p-8"
-      >
-        <LastFmRecent client:load />
-      </div>
-
-      <div
-        class="animate-fade rounded-xl bg-sweater-10 [--delay:500ms] xl:col-span-4 xl:row-span-2 xl:flex xl:h-[244px] xl:w-[500px] xl:flex-col xl:justify-between xl:gap-3 xl:bg-sweater-9 xl:p-8"
-      >
-        <div class="mb-4 xl:mb-0">
-          <h4 class="mb-4 w-fit font-bold text-sweater-2 xl:mb-0 xl:text-sm">
-            My Last.fm profile
-          </h4>
-          <p>
-            For fun and out of curiosity, I started tracking my music habits a
-            few years ago.
-          </p>
-        </div>
-        <LastFmStats client:load />
-        <LastFmWeekly client:load />
-      </div>
+      <h3 class="font-bold text-sweater-2">About</h3>
+      <p>
+        Hi, my name is Charles. I'm a student at the University of Pennsylvania
+        majoring in
+        <Link
+          href="http://cg.cis.upenn.edu/dmd.html"
+          newTab
+        >
+          <>Digital Media Design</>
+        </Link>, an interdisciplinary program that combines computer science and
+        fine arts.
+      </p>
+      <p>
+        I want to make games and build websites. To that end, I help run <Link
+          href="https://pennupgrade.com"
+          newTab
+        >
+          <>UPGRADE</>
+        </Link>, Penn's game development club. I'm also a board member and
+        webmaster for <Link
+          href="http://cg.cis.upenn.edu/siggraph/"
+          newTab
+        >
+          <>Penn's SIGGRAPH chapter</>
+        </Link>.
+      </p>
+      <p>
+        In my free time I like running, designing album covers, obsessing over
+        rail systems, practicing bass guitar, and ricing my Arch setup.
+      </p>
+      <p>Thanks for stopping by. (〜￣▽￣)〜</p>
     </section>
-  </main>
-</HtmlLayout>
+
+    <section
+      class="max-w-content mx-auto animate-fade space-y-[1.15rem] [--delay:300ms]"
+    >
+      <h3 class="mb-[1.15rem] font-bold text-sweater-2">Current status</h3>
+      <LastFm client:load />
+    </section>
+
+    <section
+      class="max-w-content mx-auto animate-fade space-y-[1.15rem] [--delay:5s00ms]"
+    >
+      <h3 class="mb-[1.15rem] font-bold text-sweater-2">Projects</h3>
+    </section>
+  </MainLayout>
+</PrimitiveLayout>
