+++ src/styles/base.css
@@ -3,17 +3,6 @@
 @tailwind utilities;
 
 @layer base {
-  h1,
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    @apply font-mono;
-  }
-}
-
-@layer base {
   img {
     @apply italic;
   }
